requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
openai>=1.3.0
python-dotenv>=1.0.0
schedule>=1.2.0
psutil>=5.9.0
colorama>=0.4.0
pydantic>=2.5.0
typing-extensions>=4.8.0
pillow>=10.0.0
customtkinter>=5.2.0
google-generativeai>=0.3.0
anthropic>=0.7.0
aiofiles>=23.0.0